requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
pydantic>=2.7.0
tavily-python>=0.5.0
python-dotenv>=1.0.1
